requirements.txt
kafka-python
faker
psycopg2-binary
pandas
plotly
streamlit
sqlalchemy
